Projects/04_LLMs/01_LLm_powered_QA_chatbot/pyproject.toml
[project]
name = "01-llm-powered-qa-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.6",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
]
